Move the contents of `wallet_utils.py` to `bittensor/utils/__init__.py`, since it is used directly from utils.
