+++ bittensor/utils/__init__.py
@@ -16,15 +16,17 @@
 # DEALINGS IN THE SOFTWARE.
 
 import hashlib
-from typing import Callable, List, Dict, Literal, Tuple
+from typing import Callable, List, Dict, Literal, Tuple, Union, Optional
 
 import numpy as np
 import scalecodec
+from substrateinterface import Keypair as Keypair
+from substrateinterface.utils import ss58
 
+from bittensor.core.settings import ss58_format
 from .registration import torch, use_torch
 from .version import version_checking, check_version, VersionCheckError
-from .wallet_utils import *  # noqa F401
-from ..core.settings import ss58_format
+
 
 RAOPERTAO = 1e9
 U16_MAX = 65535
@@ -48,26 +50,17 @@ def _unbiased_topk(
 ) -> Union[Tuple[np.ndarray, np.ndarray], Tuple["torch.Tensor", "torch.LongTensor"]]:
     """Selects topk as in torch.topk but does not bias lower indices when values are equal.
     Args:
-        values: (np.ndarray) if using numpy, (torch.Tensor) if using torch:
-            Values to index into.
-        k: (int):
-            Number to take.
-        dim: (int):
-            Dimension to index into (used by Torch)
-        sorted: (bool):
-            Whether to sort indices.
-        largest: (bool):
-            Whether to take the largest value.
-        axis: (int):
-            Axis along which to index into (used by Numpy)
-        return_type: (str):
-            Whether or use torch or numpy approach
+        values: (np.ndarray) if using numpy, (torch.Tensor) if using torch: Values to index into.
+        k (int): Number to take.
+        dim (int): Dimension to index into (used by Torch)
+        sorted (bool): Whether to sort indices.
+        largest (bool): Whether to take the largest value.
+        axis (int): Axis along which to index into (used by Numpy)
+        return_type (str): Whether or use torch or numpy approach
 
     Return:
-        topk: (np.ndarray) if using numpy, (torch.Tensor) if using torch:
-            topk k values.
-        indices: (np.ndarray) if using numpy, (torch.LongTensor) if using torch:
-            indices of the topk values.
+        topk (np.ndarray): if using numpy, (torch.Tensor) if using torch: topk k values.
+        indices (np.ndarray): if using numpy, (torch.LongTensor) if using torch: indices of the topk values.
     """
     if return_type == "torch":
         permutation = torch.randperm(values.shape[dim])
@@ -279,3 +272,142 @@ def format_error_message(error_message: dict) -> str:
         err_docs = error_message.get("docs", [])
         err_description = err_docs[0] if len(err_docs) > 0 else err_description
     return f"Subtensor returned `{err_name} ({err_type})` error. This means: `{err_description}`"
+
+
+def create_identity_dict(
+    display: str = "",
+    legal: str = "",
+    web: str = "",
+    riot: str = "",
+    email: str = "",
+    pgp_fingerprint: Optional[str] = None,
+    image: str = "",
+    info: str = "",
+    twitter: str = "",
+) -> dict:
+    """
+    Creates a dictionary with structure for identity extrinsic. Must fit within 64 bits.
+
+    Args:
+        display (str): String to be converted and stored under 'display'.
+        legal (str): String to be converted and stored under 'legal'.
+        web (str): String to be converted and stored under 'web'.
+        riot (str): String to be converted and stored under 'riot'.
+        email (str): String to be converted and stored under 'email'.
+        pgp_fingerprint (str): String to be converted and stored under 'pgp_fingerprint'.
+        image (str): String to be converted and stored under 'image'.
+        info (str): String to be converted and stored under 'info'.
+        twitter (str): String to be converted and stored under 'twitter'.
+
+    Returns:
+        dict: A dictionary with the specified structure and byte string conversions.
+
+    Raises:
+    ValueError: If pgp_fingerprint is not exactly 20 bytes long when encoded.
+    """
+    if pgp_fingerprint and len(pgp_fingerprint.encode()) != 20:
+        raise ValueError("pgp_fingerprint must be exactly 20 bytes long when encoded")
+
+    return {
+        "info": {
+            "additional": [[]],
+            "display": {f"Raw{len(display.encode())}": display.encode()},
+            "legal": {f"Raw{len(legal.encode())}": legal.encode()},
+            "web": {f"Raw{len(web.encode())}": web.encode()},
+            "riot": {f"Raw{len(riot.encode())}": riot.encode()},
+            "email": {f"Raw{len(email.encode())}": email.encode()},
+            "pgp_fingerprint": pgp_fingerprint.encode() if pgp_fingerprint else None,
+            "image": {f"Raw{len(image.encode())}": image.encode()},
+            "info": {f"Raw{len(info.encode())}": info.encode()},
+            "twitter": {f"Raw{len(twitter.encode())}": twitter.encode()},
+        }
+    }
+
+
+def decode_hex_identity_dict(info_dictionary):
+    for key, value in info_dictionary.items():
+        if isinstance(value, dict):
+            item = list(value.values())[0]
+            if isinstance(item, str) and item.startswith("0x"):
+                try:
+                    info_dictionary[key] = bytes.fromhex(item[2:]).decode()
+                except UnicodeDecodeError:
+                    print(f"Could not decode: {key}: {item}")
+            else:
+                info_dictionary[key] = item
+    return info_dictionary
+
+
+def is_valid_ss58_address(address: str) -> bool:
+    """
+    Checks if the given address is a valid ss58 address.
+
+    Args:
+        address(str): The address to check.
+
+    Returns:
+        True if the address is a valid ss58 address for Bittensor, False otherwise.
+    """
+    try:
+        return ss58.is_valid_ss58_address(
+            address, valid_ss58_format=ss58_format
+        ) or ss58.is_valid_ss58_address(
+            address, valid_ss58_format=42
+        )  # Default substrate ss58 format (legacy)
+    except IndexError:
+        return False
+
+
+def _is_valid_ed25519_pubkey(public_key: Union[str, bytes]) -> bool:
+    """
+    Checks if the given public_key is a valid ed25519 key.
+
+    Args:
+        public_key(Union[str, bytes]): The public_key to check.
+
+    Returns:
+        True if the public_key is a valid ed25519 key, False otherwise.
+
+    """
+    try:
+        if isinstance(public_key, str):
+            if len(public_key) != 64 and len(public_key) != 66:
+                raise ValueError("a public_key should be 64 or 66 characters")
+        elif isinstance(public_key, bytes):
+            if len(public_key) != 32:
+                raise ValueError("a public_key should be 32 bytes")
+        else:
+            raise ValueError("public_key must be a string or bytes")
+
+        keypair = Keypair(public_key=public_key, ss58_format=ss58_format)
+
+        ss58_addr = keypair.ss58_address
+        return ss58_addr is not None
+
+    except (ValueError, IndexError):
+        return False
+
+
+def is_valid_bittensor_address_or_public_key(address: Union[str, bytes]) -> bool:
+    """
+    Checks if the given address is a valid destination address.
+
+    Args:
+        address(Union[str, bytes]): The address to check.
+
+    Returns:
+        True if the address is a valid destination address, False otherwise.
+    """
+    if isinstance(address, str):
+        # Check if ed25519
+        if address.startswith("0x"):
+            return _is_valid_ed25519_pubkey(address)
+        else:
+            # Assume ss58 address
+            return is_valid_ss58_address(address)
+    elif isinstance(address, bytes):
+        # Check if ed25519
+        return _is_valid_ed25519_pubkey(address)
+    else:
+        # Invalid address type
+        return False
